refactor(core): move MainDispatcherRule to core:testing module
