+++ core/testing/build.gradle.kts
@@ -9,6 +9,9 @@ android {
 kotlin {
     sourceSets {
         commonMain.dependencies {
+            implementation(libs.kotlin.test)
+            implementation(libs.kotlin.test.junit)
+            implementation(libs.kotlinx.coroutines.test)
             implementation(libs.assertk)
             implementation(libs.turbine)
         }
