+++ feature/account/server/settings/src/test/kotlin/app/k9mail/feature/account/server/settings/ui/outgoing/OutgoingServerSettingsViewModelTest.kt
@@ -1,6 +1,5 @@
 package app.k9mail.feature.account.server.settings.ui.outgoing
 
-import app.k9mail.core.ui.compose.testing.MainDispatcherRule
 import app.k9mail.core.ui.compose.testing.mvi.assertThatAndMviTurbinesConsumed
 import app.k9mail.core.ui.compose.testing.mvi.eventStateTest
 import app.k9mail.core.ui.compose.testing.mvi.runMviTest
@@ -24,6 +23,7 @@ import com.fsck.k9.mail.AuthType
 import com.fsck.k9.mail.ServerSettings
 import net.thunderbird.core.common.domain.usecase.validation.ValidationError
 import net.thunderbird.core.common.domain.usecase.validation.ValidationResult
+import net.thunderbird.core.testing.coroutines.MainDispatcherRule
 import org.junit.Rule
 import org.junit.Test
