+++ feature/account/setup/src/test/kotlin/app/k9mail/feature/account/setup/ui/options/sync/SyncOptionsViewModelTest.kt
@@ -1,6 +1,5 @@
 package app.k9mail.feature.account.setup.ui.options.sync
 
-import app.k9mail.core.ui.compose.testing.MainDispatcherRule
 import app.k9mail.core.ui.compose.testing.mvi.assertThatAndEffectTurbineConsumed
 import app.k9mail.core.ui.compose.testing.mvi.eventStateTest
 import app.k9mail.core.ui.compose.testing.mvi.runMviTest
@@ -15,6 +14,7 @@ import app.k9mail.feature.account.setup.ui.options.sync.SyncOptionsContract.Even
 import app.k9mail.feature.account.setup.ui.options.sync.SyncOptionsContract.State
 import assertk.assertThat
 import assertk.assertions.isEqualTo
+import net.thunderbird.core.testing.coroutines.MainDispatcherRule
 import org.junit.Rule
 import org.junit.Test
