+++ feature/account/setup/src/test/kotlin/app/k9mail/feature/account/setup/ui/specialfolders/SpecialFoldersViewModelTest.kt
@@ -1,6 +1,5 @@
 package app.k9mail.feature.account.setup.ui.specialfolders
 
-import app.k9mail.core.ui.compose.testing.MainDispatcherRule
 import app.k9mail.core.ui.compose.testing.mvi.assertThatAndEffectTurbineConsumed
 import app.k9mail.core.ui.compose.testing.mvi.assertThatAndStateTurbineConsumed
 import app.k9mail.core.ui.compose.testing.mvi.runMviTest
@@ -26,6 +25,7 @@ import com.fsck.k9.mail.folders.RemoteFolder
 import kotlinx.coroutines.delay
 import net.thunderbird.core.common.domain.usecase.validation.ValidationError
 import net.thunderbird.core.common.domain.usecase.validation.ValidationResult
+import net.thunderbird.core.testing.coroutines.MainDispatcherRule
 import org.junit.Rule
 import org.junit.Test
